+++ src/cdn/provision-environment.ts
@@ -14,10 +14,14 @@ async function provisionEnvironment(
   }).start()
   try {
     const body = await oneBlinkAPIClient.postRequest<
-      { isSinglePageApplication?: boolean },
+      {
+        isSinglePageApplication?: boolean
+        disableSecurityResponseHeaders?: boolean
+      },
       { brandedUrl: string }
     >(`/webApps/${cfg.scope}/environments/${env}/deployments`, {
       isSinglePageApplication: cfg.isSinglePageApplication,
+      disableSecurityResponseHeaders: cfg.disableSecurityResponseHeaders,
     })
     spinner.succeed(
       'Deployment complete - Origin: ' + chalk.bold(body.brandedUrl),